refactor: query and fragments
