+++ packages/mirror-media-next/pages/topic/[slug].js
@@ -191,7 +191,7 @@ export async function getServerSideProps({ query, req, res }) {
           RENDER_PAGE_SIZE * 2,
           posts.length
         )
-        if (!Array.isArray(topicData.data.topics[0].posts)) break
+        if (!Array.isArray(topicData.data.topics?.[0]?.posts)) break
         moreTopicPosts = topicData.data.topics[0].posts
       } catch (error) {
         logGqlError(
